+++ templates/Courses/index.php
@@ -3,16 +3,17 @@
     <div id="map"></div>
 
     <div id="map-footer" class="footer">
-        <p class="imprint"><?= $this->Html->link(
-                                'Imprint',
-                                '/pages/info/#imprint'
-                            ) ?></p>
-        <p class="license"><?= $this->Html->link(
-                                'CC-BY 4.0',
-                                'https://creativecommons.org/licenses/by/4.0/',
-                                ['target' => '_blank']
-                            ) ?></p>
-        <p class="copyright">&copy;2014-<?= date('Y') ?></p>
+        <p class="imprint">
+            <?= $this->Html->link('Imprint', '/pages/info/#imprint') ?>
+            &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;
+            <?= $this->Html->link('FAQ', ['controller' => 'faqQuestions', 'action' => 'faqList', 'public']) ?>
+        </p>
+        <p class="license">
+            <?= $this->Html->link('CC-BY 4.0', 'https://creativecommons.org/licenses/by/4.0/', ['target' => '_blank']) ?>
+        </p>
+        <p class="copyright">
+            &copy;2014-<?= date('Y') ?>
+        </p>
     </div>
 </div>
 <div id="slide-control"></div>