+++ templates/element/info/national-moderators.php
@@ -1,0 +1,80 @@
+<style>
+    .card {
+        -moz-border-radius: 2%;
+        -webkit-border-radius: 2%;
+        border-radius: 2%;
+        box-shadow: 5px 5px 0 rgba(0, 0, 0, 0.08);
+    }
+
+    .profile .profile-body {
+        padding: 20px;
+        background: #f7f7f7;
+    }
+
+    .profile .profile-bio {
+        background: #fff;
+        position: relative;
+        padding: 15px 10px 5px 15px;
+    }
+
+    .profile .profile-bio a {
+        left: 50%;
+        bottom: 20px;
+        margin: -62px;
+        text-align: center;
+        position: absolute;
+    }
+
+    .profile .profile-bio h2 {
+        margin-top: 0;
+        font-weight: 200;
+    }
+</style>
+
+<p>The DHCR maintains a system of national moderators, who review newly entered courses, help with registration issues and are encouraged 
+to disseminate the DHCR initiative among institutions of their countries.</p>
+
+<div class="container">
+    <div class="profile card">
+        <div class="profile-body">
+            <div class="profile-bio">
+                <h3>Country1</h3>
+                <div class="row">
+                    <div class="col-md-5 text-center">
+                        <img class="img-thumbnail md-margin-bottom-10" src="img/map_mobil.png" height="200" alt="">
+                    </div>
+                    <div class="col-md-7">
+                        <h2>Firstname Lastname</h2>
+                        <p><strong>Institution</strong><br>
+                        <strong>mail at domain.ext</strong></p>
+                        <hr>
+                        <p>Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut tristique et egestas quis ipsum suspendisse ultrices. Cursus turpis massa tincidunt dui ut ornare lectus. Sed risus pretium quam vulputate. Non curabitur gravida arcu ac tortor dignissim convallis. Suscipit adipiscing bibendum est ultricies integer quis auctor elit. Porta non pulvinar neque laoreet suspendisse interdum consectetur. Sit amet aliquam id diam maecenas. Non pulvinar neque laoreet suspendisse interdum. Sed cras ornare arcu dui vivamus arcu. Amet mattis vulputate enim nulla aliquet porttitor lacus luctus accumsan. Lobortis feugiat vivamus at augue eget arcu dictum varius. Laoreet suspendisse interdum consectetur libero id faucibus nisl. Eros in cursus turpis massa. Tortor condimentum lacinia quis vel eros donec ac odio.</p>
+                    </div>
+                </div>
+            </div>
+        </div>
+    </div>
+</div>
+<p>&nbsp;</p>
+<div class="container">
+    <div class="profile card">
+        <div class="profile-body">
+            <div class="profile-bio">
+                <h3>Country2</h3>
+                <div class="row">
+                    <div class="col-md-5 text-center">
+                        <img class="img-thumbnail md-margin-bottom-10" src="img/map_mobil.png" height="200" alt="">
+                    </div>
+                    <div class="col-md-7">
+                        <h2>Firstname Lastname</h2>
+                        <p><strong>Institution</strong><br>
+                        <strong>mail at domain.ext</strong></p>
+                        <hr>
+                        <p>Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut tristique et egestas quis ipsum suspendisse ultrices. Cursus turpis massa tincidunt dui ut ornare lectus. Sed risus pretium quam vulputate. Non curabitur gravida arcu ac tortor dignissim convallis. Suscipit adipiscing bibendum est ultricies integer quis auctor elit. Porta non pulvinar neque laoreet suspendisse interdum consectetur. Sit amet aliquam id diam maecenas. Non pulvinar neque laoreet suspendisse interdum. Sed cras ornare arcu dui vivamus arcu. Amet mattis vulputate enim nulla aliquet porttitor lacus luctus accumsan. Lobortis feugiat vivamus at augue eget arcu dictum varius. Laoreet suspendisse interdum consectetur libero id faucibus nisl. Eros in cursus turpis massa. Tortor condimentum lacinia quis vel eros donec ac odio.</p>
+                    </div>
+                </div>
+            </div>
+        </div>
+    </div>
+</div>
+<p>&nbsp;</p>