+++ templates/element/info/social-media.php
@@ -1,0 +1,8 @@
+<div class="title">
+    <h2>Follow the DH Course Registry</h2>
+</div>
+<?= $this->Element('follow/social_media'); ?>
+<p>Click
+    <?= $this->Html->link('here', ['controller' => 'Pages', 'action' => 'follow']) ?>
+    the to learn how you can stay informed about new courses or the latest technical enhancements.
+</p>