+++ templates/element/sitemap.php
@@ -30,6 +30,16 @@
                         ['controller' => 'Pages', 'action' => 'info', '#' => 'contact']
                     ) ?>
                 </li>
+                <?php
+                /*
+                <li>
+                    <?= $this->Html->link(
+                        'National Moderators',
+                        ['controller' => 'Pages', 'action' => 'info', '#' => 'national-moderators']
+                    ) ?>
+                </li>
+                */
+                ?>
                 <li>
                     <?= $this->Html->link(
                         'Dissemination and Impact',
@@ -46,6 +56,12 @@
                     <?= $this->Html->link(
                         'CLARIN and DARIAH',
                         ['controller' => 'Pages', 'action' => 'info', '#' => 'clarin-dariah']
+                    ) ?>
+                </li>
+                <li>
+                    <?= $this->Html->link(
+                        'FAQ',
+                        ['controller' => 'faqQuestions', 'action' => 'faqList', 'public']
                     ) ?>
                 </li>
                 <li>